Detect an excursion in many timeseries

Determines whether an excursion event has occurred within the specified event window for a lipd-ts-tibble of timeseries. Excursion events are defined as n.consecutive values within the event window that are more extreme than the avg +/- sig.num standard deviations of the reference windows.

detectExcursionSlidingWindow(
  ltt = NA,
  time = NA,
  vals = NA,
  time.variable.name = NA,
  vals.variable.name = NA,
  time.units = NA,
  vals.units = NA,
  dataset.name = NA,
  event.yr.vec = NA,
  event.step = NA,
  ref.window,
  ...
)

Arguments

ltt

A LiPD-timeseries-tibble, a tibble or data.frame that has the variable(s) of interest, a time variable (age, year or time) along with their metadata, aranged in rows. If ltt = NA, then one in is created from other inputs

time

if ltt is not provided, input a vector or matrix of time (year or age) data. If it's a multicolumn matrix, the columns are time-ensemble members

vals

if ltt is not provided, input a vector or matrix of paleoData. If it's a multicolumn matrix, the columns are value-ensemble members

time.variable.name

If ltt is not provided, specify the name of the time variable (typically 'age' or 'year')

vals.variable.name

If ltt is not provided, specify the name of the paleo variable (e.g., 'd18O' or 'temperature'). Alternatively, if ltt is provided with more rows than expected, this term is used to attempt to select the correct row.

time.units

If ltt is not provided, specify the units the time variable (typically 'yr BP' or 'CE')

vals.units

If ltt is not provided, specify the units the paleo variable (e.g. 'permil' or 'degrees C')

dataset.name

If ltt is not provided, specify the dataset name

event.yr.vec

A vector of times at the center of event years to test.

event.step

If event.yr.vec = NA, then event.step will build one with this spacing.

ref.window

width (in time units) of the reference windows

...

arguments to pass to pass to changeFun

seed

Set a seed for reproducibility. By default it will use current time meaning it will not be reproducible.

Value

a tibble that describes the positive and negative excursion results

References

Morrill

Author

Hannah Kolus

Nick McKay