Prepare input for actR detection methods

Prepare input for actR detection methods

prepareInput(
  ltt = NA,
  time = NA,
  vals = NA,
  time.variable.name = NA,
  vals.variable.name = NA,
  time.units = NA,
  vals.units = NA,
  dataset.name = NA,
  expecting.one.row = TRUE,
  sort.by.time = TRUE,
  remove.time.nas = TRUE,
  time.range = NA
)

Arguments

ltt

A LiPD-timeseries-tibble, a tibble or data.frame that has the variable(s) of interest, a time variable (age, year or time) along with their metadata, aranged in rows. If ltt = NA, then one in is created from other inputs

time

if ltt is not provided, input a vector or matrix of time (year or age) data. If it's a multicolumn matrix, the columns are time-ensemble members

vals

if ltt is not provided, input a vector or matrix of paleoData. If it's a multicolumn matrix, the columns are value-ensemble members

time.variable.name

If ltt is not provided, specify the name of the time variable (typically 'age' or 'year')

vals.variable.name

If ltt is not provided, specify the name of the paleo variable (e.g., 'd18O' or 'temperature'). Alternatively, if ltt is provided with more rows than expected, this term is used to attempt to select the correct row.

time.units

If ltt is not provided, specify the units the time variable (typically 'yr BP' or 'CE')

vals.units

If ltt is not provided, specify the units the paleo variable (e.g. 'permil' or 'degrees C')

dataset.name

If ltt is not provided, specify the dataset name

expecting.one.row

Are you expecting a one-row tibble? That is, one variable of interest? (default = TRUE)

sort.by.time

Do you want to arrange the output so that the time variable is increasing, while preserving the time-value relationship? (default = TRUE)

remove.time.nas

Do you want to remove observations that are NA, or otherwise not finite, in the time variable (default = TRUE)

time.range

Optionally enter a time range (as minimum and maximum) that you'd like to restrict the analysis to. (default = NA)

Value

a lipd-ts-tibble ready for analysis in actR