Explore uncertainty space on abscissa or ordinate and propagate of any of actR's change functions

Explore uncertainty space on abscissa or ordinate and propagate of any of actR's change functions

propagateUncertainty(
  time,
  vals,
  changeFun,
  simulate.time.uncertainty = TRUE,
  simulate.paleo.uncertainty = TRUE,
  n.ens = 100,
  bam.model = list(ns = n.ens, name = "bernoulli", param = 0.05),
  paleo.uncertainty = sd(vals, na.rm = TRUE)/2,
  paleo.ar1 = sqrt(0.5),
  paleo.arima.order = c(1, 0, 0),
  summarize = FALSE,
  seed = round(sum(time, na.rm = TRUE)),
  progress = TRUE,
  ...
)

Arguments

time

a time vector, or matrix of time ensemble members (ensembles in columns)

vals

a values vector, or matrix of values ensemble members (ensembles in columns)

changeFun

the change function to across which to propagate

simulate.time.uncertainty

TRUE or FALSE. If an ensemble is not included, do you want to simulate time ensembles (default = TRUE)

simulate.paleo.uncertainty

TRUE or FALSE. If an ensemble is not included, do you want to simulate paleo ensembles (default = TRUE)

n.ens

How many ensembles to use for error propagation? (default = 100)

bam.model

BAM Model parameters to use if simulating time uncertainty (default = list(ns = n.ens, name = "bernoulli", param = 0.05), paleo.uncertainty = sd(vals,na.rm = TRUE)))

paleo.uncertainty

Uncertainty to use if modelling uncertainty for paleo values. (default = sd(vals,na.rm = TRUE)/2)

paleo.ar1

Autocorrelation coefficient to use for modelling uncertainty on paleoData, what fraction of the uncertainties are autocorrelated? (default = sqrt(0.5); or 50% autocorrelated uncertainty)

paleo.arima.order

Order to use for ARIMA model used in modelling uncertainty on paleoDat (default = c(1,0,0))

summarize

Boolean. Summarize the output? Or return all the ensembles?

seed

set a seed for reproducibility

progress

show null hypothesis testing progress bar?

...

arguments to pass to pass to changeFun

Value

a propagated uncertainty tibble